Use the proper array operator (Visual Studio was confused between two possible alternatives)
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@26898 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/spds/MailMessage.cpp
@@ -204,7 +204,7 @@ inline static size_t findNewLine(StringBuffer &str, size_t offset) {
     size_t nl = str.find("\n", offset)+1;
     if(nl == StringBuffer::npos)
         return nl;
-    return (str[nl] == '\r') ? nl+1 : nl ;
+    return (str[(int)nl] == '\r') ? nl+1 : nl ;
 }
 
 static size_t getHeadersLen(StringBuffer &s, StringBuffer &newline)
@@ -217,7 +217,7 @@ static size_t getHeadersLen(StringBuffer &s, StringBuffer &newline)
     }
     size_t pos2 = pos1 + 1 ;
 
-    while (s[pos1-1] == '\r'){
+    while (s[(int)(pos1-1)] == '\r'){
         pos1--;
     }
     newline = s.substr(pos1, pos2-pos1);
